Move `Rounding` to its own file
