+++ crates/epaint/src/lib.rs
@@ -29,6 +29,7 @@ pub mod image;
 mod margin;
 mod mesh;
 pub mod mutex;
+mod rounding;
 mod shadow;
 mod shape;
 pub mod shape_transform;
@@ -47,10 +48,11 @@ pub use self::{
     image::{ColorImage, FontImage, ImageData, ImageDelta},
     margin::Margin,
     mesh::{Mesh, Mesh16, Vertex},
+    rounding::Rounding,
     shadow::Shadow,
     shape::{
-        CircleShape, EllipseShape, PaintCallback, PaintCallbackInfo, PathShape, RectShape,
-        Rounding, Shape, TextShape,
+        CircleShape, EllipseShape, PaintCallback, PaintCallbackInfo, PathShape, RectShape, Shape,
+        TextShape,
     },
     stats::PaintStats,
     stroke::{PathStroke, Stroke, StrokeKind},
