+++ crates/epaint/src/shape.rs
@@ -5,7 +5,7 @@ use std::{any::Any, sync::Arc};
 use crate::{
     stroke::PathStroke,
     text::{FontId, Fonts, Galley},
-    Color32, Mesh, Stroke, TextureId,
+    Color32, Mesh, Rounding, Stroke, TextureId,
 };
 use emath::{pos2, Align2, Pos2, Rangef, Rect, TSTransform, Vec2};
 
@@ -776,214 +776,6 @@ impl From<RectShape> for Shape {
     #[inline(always)]
     fn from(shape: RectShape) -> Self {
         Self::Rect(shape)
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-/// How rounded the corners of things should be
-pub struct Rounding {
-    /// Radius of the rounding of the North-West (left top) corner.
-    pub nw: f32,
-
-    /// Radius of the rounding of the North-East (right top) corner.
-    pub ne: f32,
-
-    /// Radius of the rounding of the South-West (left bottom) corner.
-    pub sw: f32,
-
-    /// Radius of the rounding of the South-East (right bottom) corner.
-    pub se: f32,
-}
-
-impl Default for Rounding {
-    #[inline]
-    fn default() -> Self {
-        Self::ZERO
-    }
-}
-
-impl From<f32> for Rounding {
-    #[inline]
-    fn from(radius: f32) -> Self {
-        Self {
-            nw: radius,
-            ne: radius,
-            sw: radius,
-            se: radius,
-        }
-    }
-}
-
-impl Rounding {
-    /// No rounding on any corner.
-    pub const ZERO: Self = Self {
-        nw: 0.0,
-        ne: 0.0,
-        sw: 0.0,
-        se: 0.0,
-    };
-
-    #[inline]
-    pub const fn same(radius: f32) -> Self {
-        Self {
-            nw: radius,
-            ne: radius,
-            sw: radius,
-            se: radius,
-        }
-    }
-
-    /// Do all corners have the same rounding?
-    #[inline]
-    pub fn is_same(&self) -> bool {
-        self.nw == self.ne && self.nw == self.sw && self.nw == self.se
-    }
-
-    /// Make sure each corner has a rounding of at least this.
-    #[inline]
-    pub fn at_least(&self, min: f32) -> Self {
-        Self {
-            nw: self.nw.max(min),
-            ne: self.ne.max(min),
-            sw: self.sw.max(min),
-            se: self.se.max(min),
-        }
-    }
-
-    /// Make sure each corner has a rounding of at most this.
-    #[inline]
-    pub fn at_most(&self, max: f32) -> Self {
-        Self {
-            nw: self.nw.min(max),
-            ne: self.ne.min(max),
-            sw: self.sw.min(max),
-            se: self.se.min(max),
-        }
-    }
-}
-
-impl std::ops::Add for Rounding {
-    type Output = Self;
-    #[inline]
-    fn add(self, rhs: Self) -> Self {
-        Self {
-            nw: self.nw + rhs.nw,
-            ne: self.ne + rhs.ne,
-            sw: self.sw + rhs.sw,
-            se: self.se + rhs.se,
-        }
-    }
-}
-
-impl std::ops::AddAssign for Rounding {
-    #[inline]
-    fn add_assign(&mut self, rhs: Self) {
-        *self = Self {
-            nw: self.nw + rhs.nw,
-            ne: self.ne + rhs.ne,
-            sw: self.sw + rhs.sw,
-            se: self.se + rhs.se,
-        };
-    }
-}
-
-impl std::ops::AddAssign<f32> for Rounding {
-    #[inline]
-    fn add_assign(&mut self, rhs: f32) {
-        *self = Self {
-            nw: self.nw + rhs,
-            ne: self.ne + rhs,
-            sw: self.sw + rhs,
-            se: self.se + rhs,
-        };
-    }
-}
-
-impl std::ops::Sub for Rounding {
-    type Output = Self;
-    #[inline]
-    fn sub(self, rhs: Self) -> Self {
-        Self {
-            nw: self.nw - rhs.nw,
-            ne: self.ne - rhs.ne,
-            sw: self.sw - rhs.sw,
-            se: self.se - rhs.se,
-        }
-    }
-}
-
-impl std::ops::SubAssign for Rounding {
-    #[inline]
-    fn sub_assign(&mut self, rhs: Self) {
-        *self = Self {
-            nw: self.nw - rhs.nw,
-            ne: self.ne - rhs.ne,
-            sw: self.sw - rhs.sw,
-            se: self.se - rhs.se,
-        };
-    }
-}
-
-impl std::ops::SubAssign<f32> for Rounding {
-    #[inline]
-    fn sub_assign(&mut self, rhs: f32) {
-        *self = Self {
-            nw: self.nw - rhs,
-            ne: self.ne - rhs,
-            sw: self.sw - rhs,
-            se: self.se - rhs,
-        };
-    }
-}
-
-impl std::ops::Div<f32> for Rounding {
-    type Output = Self;
-    #[inline]
-    fn div(self, rhs: f32) -> Self {
-        Self {
-            nw: self.nw / rhs,
-            ne: self.ne / rhs,
-            sw: self.sw / rhs,
-            se: self.se / rhs,
-        }
-    }
-}
-
-impl std::ops::DivAssign<f32> for Rounding {
-    #[inline]
-    fn div_assign(&mut self, rhs: f32) {
-        *self = Self {
-            nw: self.nw / rhs,
-            ne: self.ne / rhs,
-            sw: self.sw / rhs,
-            se: self.se / rhs,
-        };
-    }
-}
-
-impl std::ops::Mul<f32> for Rounding {
-    type Output = Self;
-    #[inline]
-    fn mul(self, rhs: f32) -> Self {
-        Self {
-            nw: self.nw * rhs,
-            ne: self.ne * rhs,
-            sw: self.sw * rhs,
-            se: self.se * rhs,
-        }
-    }
-}
-
-impl std::ops::MulAssign<f32> for Rounding {
-    #[inline]
-    fn mul_assign(&mut self, rhs: f32) {
-        *self = Self {
-            nw: self.nw * rhs,
-            ne: self.ne * rhs,
-            sw: self.sw * rhs,
-            se: self.se * rhs,
-        };
     }
 }
